# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Nina Ofertas API",
  "version": "0.5.0"
 },
 "paths": {
...
  "/api/bots": {
   "get": {
    "tags": [
     "bots"
    ],
    "summary": "List Bots",
    "operationId": "list_bots_api_bots_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "niche_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Niche Id"
      }
     },
     {
      "name": "phone_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Phone Id"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "-created_at",
       "title": "Sort"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PaginatedResponse_BotResponse_"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "bots"
    ],
    "summary": "Create Bot",
    "operationId": "create_bot_api_bots_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BotCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BotResponse"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   }
  },
  "/api/bots/{bot_id}": {
   "get": {
    "tags": [
     "bots"
    ],
    "summary": "Get Bot",
    "operationId": "get_bot_api_bots__bot_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "bot_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Bot Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BotResponse"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "bots"
    ],
    "summary": "Update Bot",
    "operationId": "update_bot_api_bots__bot_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "bot_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Bot Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BotUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BotResponse"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "bots"
    ],
    "summary": "Delete Bot",
    "operationId": "delete_bot_api_bots__bot_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "bot_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Bot Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BotCreate": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Name"
     },
     "slug": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
      "title": "Slug"
     },
     "niche_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Niche Id"
     },
     "phone_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone Id"
     },
     "settings": {
      "$ref": "#/components/schemas/BotSettings"
     },
     "message_template": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Message Template"
     },
     "group_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "title": "Group Ids"
     },
     "account_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "title": "Account Ids"
     }
    },
    "type": "object",
    "required": [
     "name",
     "slug",
     "settings"
    ],
    "title": "BotCreate"
   },
...
   "BotUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "slug": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100,
        "minLength": 1,
        "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Slug"
     },
     "niche_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Niche Id"
     },
     "phone_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone Id"
     },
     "settings": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/BotSettings"
       },
       {
        "type": "null"
       }
      ]
     },
     "message_template": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Message Template"
     },
     "status": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "active",
         "paused",
         "disabled"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Status"
     }
    },
    "type": "object",
    "title": "BotUpdate"
   },
...
   "ErrorEnvelope": {
    "properties": {
     "error": {
      "$ref": "#/components/schemas/ErrorDetail"
     }
    },
    "type": "object",
    "required": [
     "error"
    ],
    "title": "ErrorEnvelope"
   },
...
   "PaginatedResponse_BotResponse_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/BotResponse"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "PaginatedResponse[BotResponse]"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}